1. Deleted auth.py from auth 2. Created default roles (owner, barber, client) 3. Created test users for each role with the same name and password:password 4. Users will be given the role client by default when created 5. Expanded admin permissions on barbershop realm 6. Added Keycloak registration route to user_router.py
